+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 1.1.9: October 17th, 2018
 * **Tweak** - Refactoring post/body classes
 * **Tweak** - Package price with option to select accommodation, for best or without accommodation
 * **Tweak** - Package field "package available" not required